+++ app/src/main/java/cityfreqs/com/pilfershushjammer/AudioChecker.java
@@ -33,7 +33,7 @@ public class AudioChecker {
                         AudioFormat.CHANNEL_IN_MONO,    // 16, also CHANNEL_IN_FRONT == 16
                         AudioFormat.CHANNEL_IN_STEREO }) {  // 12
                     try {
-                        MainActivity.entryLogger("Try rate " + rate + "Hz, bits: " + audioFormat + ", channelInConfig: "+ channelInConfig, false);
+                        if (DEBUG) MainActivity.entryLogger("Try rate " + rate + "Hz, bits: " + audioFormat + ", channelInConfig: "+ channelInConfig, false);
                         buffSize = AudioRecord.getMinBufferSize(rate, channelInConfig, audioFormat);
                         // force buffSize to powersOfTwo if it isnt (ie.S5)
                         buffSize = AudioSettings.getClosestPowersHigh(buffSize);
@@ -47,10 +47,10 @@ public class AudioChecker {
                                     buffSize);
 
                             if (recorder.getState() == AudioRecord.STATE_INITIALIZED) {
-                                MainActivity.entryLogger("found: " + rate + ", buffer: " + buffSize + ", channel count: " + recorder.getChannelCount(), true);
-                                // set found values
                                 // AudioRecord.getChannelCount() is number of input audio channels (1 is mono, 2 is stereo)
-                                audioSettings.setBasicAudioInSettings(rate, buffSize, audioFormat, channelInConfig, recorder.getChannelCount());
+                                if (DEBUG) MainActivity.entryLogger("found: " + rate + ", buffer: " + buffSize + ", channel count: " + recorder.getChannelCount(), true);
+                                // set found values
+                                audioSettings.setBasicAudioInSettings(rate, buffSize, audioFormat, channelInConfig);
                                 audioSettings.setAudioSource(audioSource);
                                 recorder.release();
                                 return true;
@@ -58,12 +58,11 @@ public class AudioChecker {
                         }
                     }
                     catch (Exception e) {
-                        MainActivity.entryLogger("Error, keep trying.", false);
+                        if (DEBUG) MainActivity.entryLogger("Error, keep trying.", false);
                     }
                 }
             }
         }
-        MainActivity.entryLogger(context.getString(R.string.audio_check_1), true);
         return false;
     }
 
@@ -80,7 +79,7 @@ public class AudioChecker {
                         AudioFormat.CHANNEL_OUT_MONO,    // 4
                         AudioFormat.CHANNEL_OUT_STEREO }) {  // 12
                     try {
-                        MainActivity.entryLogger("Try rate " + rate + "Hz, bits: " + audioFormat + ", channelOutConfig: "+ channelOutConfig, false);
+                        if (DEBUG) MainActivity.entryLogger("Try rate " + rate + "Hz, bits: " + audioFormat + ", channelOutConfig: "+ channelOutConfig, false);
 
                         buffSize = AudioTrack.getMinBufferSize(rate, channelOutConfig, audioFormat);
                         // dont need to force buffSize to powersOfTwo if it isnt (ie.S5) as no FFT
@@ -95,7 +94,7 @@ public class AudioChecker {
 
 
                         if (audioTrack.getState() == AudioTrack.STATE_INITIALIZED) {
-                            MainActivity.entryLogger("found: " + rate + ", buffer: " + buffSize + ", channelOutConfig: " + channelOutConfig, true);
+                            if (DEBUG) MainActivity.entryLogger("found: " + rate + ", buffer: " + buffSize + ", channelOutConfig: " + channelOutConfig, true);
                             // set output values
                             // buffOutSize may not be same as buffInSize conformed to powersOfTwo
                             audioSettings.setChannelOutConfig(channelOutConfig);
@@ -103,11 +102,11 @@ public class AudioChecker {
 
                             // test onboardEQ
                             if (testOnboardEQ(audioTrack.getAudioSessionId())) {
-                                MainActivity.entryLogger(context.getString(R.string.eq_check_2)+ "\n", false);
+                                if (DEBUG) MainActivity.entryLogger(context.getString(R.string.eq_check_2)+ "\n", false);
                                 audioSettings.setHasEQ(true);
                             }
                             else {
-                                MainActivity.entryLogger(context.getString(R.string.eq_check_3) + "\n", true);
+                                if (DEBUG) MainActivity.entryLogger(context.getString(R.string.eq_check_3) + "\n", true);
                                 audioSettings.setHasEQ(false);
                             }
                             audioTrack.pause();
@@ -117,12 +116,11 @@ public class AudioChecker {
                         }
                     }
                     catch (Exception e) {
-                        MainActivity.entryLogger("Error, keep trying.", false);
+                        if (DEBUG) MainActivity.entryLogger("Error, keep trying.", false);
                     }
                 }
             }
         }
-        MainActivity.entryLogger(context.getString(R.string.audio_check_2), true);
         return false;
     }
 
@@ -155,7 +153,7 @@ public class AudioChecker {
                 }
             }
             // only active test is to squash all freqs in bands 0-3, leaving last band (4) free...
-            MainActivity.entryLogger("\n"+ context.getString(R.string.eq_check_7) + minEQ, false);
+            if (DEBUG) MainActivity.entryLogger("\n"+ context.getString(R.string.eq_check_7) + minEQ, false);
 
             for (int i = 0; i < 2; i++) {
                 for (short j = 0; j < bands; j++) {